NetworkMetrics to subscription
