+++ src/model/Subscribe/BlockNumber/types.ts
@@ -1,5 +1,5 @@
-// Copyright 2024 @polkadot-cloud/polkadot-developer-console authors & contributors
-// SPDX-License-Identifier: AGPL-3.0
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
 
 export interface BlockNumberEventDetail {
   blockNumber: number;
